Create show_info.py
