+++ minis/downloader/__init__.py
@@ -2,3 +2,4 @@ from .downloader import *
 from .pages import *
 from .filters import *
 from .file_utils import *
+from .show_info import *
