+++ minis/downloader/show_info.py
@@ -1,0 +1,11 @@
+
+def printDownload(currentPage, pageIndex, saveAs):
+    print(f'currentPage : {currentPage}')
+    print(f'pageIndex   : {pageIndex}')
+    print(f'saveAs      : {saveAs}')
+
+def printMiniMetadata(mini_id, name, downloadLink):
+    print(f"mini_id         : {mini_id}")
+    print(f"name            : {name}")
+    print(f"downloadLink    : {downloadLink}")
+    print(f"")
